api/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
gunicorn==23.0.0
httpx==0.28.1
redis==7.1.0
pyjwt[crypto]==2.10.1
psycopg[binary]==3.3.2
python-dotenv==1.2.1
requests==2.32.5
pandas==2.3.3
pandas-ta>=0.4.67b0
numpy>=1.26.0,<2.3
twelvedata==1.2.25
tqdm==4.67.1
websocket-client==1.9.0

